% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition}
\alias{fractionalAirlineDecomposition}
\title{Title}
\usage{
fractionalAirlineDecomposition(y, period, adjust = T, sn = F)
}
\arguments{
\item{y}{}

\item{period}{}

\item{adjust}{}

\item{sn}{}
}
\value{

}
\description{
Title
}
