% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineEstimation}
\alias{fractionalAirlineEstimation}
\title{Title}
\usage{
fractionalAirlineEstimation(
  y,
  periods,
  x = NULL,
  mean = FALSE,
  outliers = NULL,
  criticalValue = 6,
  precision = 1e-12,
  approximateHessian = F
)
}
\arguments{
\item{y}{}

\item{periods}{}

\item{x}{}

\item{mean}{}

\item{outliers}{}

\item{criticalValue}{}
}
\value{

}
\description{
Title
}
