% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{henderson}
\alias{henderson}
\title{Title}
\usage{
henderson(y, length, musgrave = TRUE, ic = 4.5)
}
\arguments{
\item{y}{}

\item{length}{}

\item{musgrave}{}

\item{ic}{}
}
\value{

}
\description{
Title
}
