% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{Holidays}
\alias{Holidays}
\title{Constructor for a holidays calendar}
\usage{
Holidays(...)
}
\arguments{
\item{...}{a serie of holiday items. One of :
\itemize{
\item a prespecified calender, see \code{\link[=listPredefinedCalendars]{listPredefinedCalendars()}}
\item a prespecified holiday, see \code{\link[=listPrespecifiedHolidays]{listPrespecifiedHolidays()}}
\item an integer vector of length 2 (month, day) indication a specific day
\item a named list containing as first item one of the above possibilities (excluding a calender) with in addition optional the following named elementd :
\itemize{
\item \emph{offset} : to indicate multiple days around specified day
\item \emph{weight} : optional a weight
\item \emph{gregorian}  : a logical
}
}}
}
\value{
\link{Holidays} a named list containing the calendar created in JDemetra. In addition it contains some labels and the used arguments to construct it.
}
\description{
Constructor for a holidays calendar
}
\examples{
\dontrun{
  # Just one day on 12th june
  Holidays(c(6,12))
  # Just one day being a prespecified day called _Christmas°
  Holidays("Christmas")
  # combining both
  Holidays(c(6,12), "Christmas")
  # combining both but also the day before and after it included
  Holidays(list(c(6,12), offset=-1:1),
           list("Christmas", offset=-1:1))
  # Reuse default calendar already specified for Belgium
  Holidays("Belgium")
  # Reuse default calendar already specified for Belgium but add some more days
  Holidays("Belgium", c(6, 12))
}
}
