% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{listPredefinedCalendars}
\alias{listPredefinedCalendars}
\title{List some predefined calendars in this package.}
\usage{
listPredefinedCalendars(onlyNames = T)
}
\description{
List some predefined calendars in this package.
}
