% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{listPrespecifiedHolidays}
\alias{listPrespecifiedHolidays}
\title{List the prespecified holidays in JDemetra}
\usage{
listPrespecifiedHolidays()
}
\description{
List the prespecified holidays in JDemetra
}
\details{
Note : should request it to JDemetra API, is currently hardcoded, copied from enum demetra.timeseries.calendars.DayEvent
}
