% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{localpolynomials}
\alias{localpolynomials}
\title{Title}
\usage{
localpolynomials(
  y,
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian",
    "Triangular", "Parabolic"),
  endpoints = c("DAF", "CC", "LC", "QL", "CQ"),
  ic = 4.5
)
}
\arguments{
\item{y}{}

\item{horizon}{}

\item{degree}{}

\item{kernel}{}

\item{endpoints}{}

\item{ic}{}
}
\value{

}
\description{
Title
}
