% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stl.R
\name{stl}
\alias{stl}
\title{Title}
\usage{
stl(y, period, multiplicative = TRUE, swindow = 7, twindow = 0, robust = TRUE)
}
\arguments{
\item{y}{}

\item{period}{}

\item{multiplicative}{}

\item{swindow}{}

\item{twindow}{}

\item{robust}{}
}
\value{

}
\description{
Title
}
