% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{x11}
\alias{x11}
\title{Title}
\usage{
x11(
  y,
  period,
  mul = TRUE,
  trend.horizon = 6,
  trend.degree = 2,
  trend.kernel = c("Henderson", "BiWeight", "TriWeight", "TriCube", "Uniform",
    "Triangular", "Epanechnikov", "Trapezoidal"),
  trend.asymmetric = c("CutAndNormalize", "Direct", "MMSRE"),
  seas.s0 = c("S3X3", "S3X1", "S3X5", "S3X9", "S3X15"),
  seas.s1 = c("S3X5", "S3X3", "S3X1", "S3X9", "S3X15"),
  extreme.lsig = 1.5,
  extreme.usig = 2.5
)
}
\arguments{
\item{y}{}

\item{period}{}

\item{mul}{}

\item{trend.horizon}{}

\item{trend.degree}{}

\item{trend.kernel}{}

\item{trend.asymmetric}{}

\item{seas.s0}{}

\item{seas.s1}{}

\item{extreme.lsig}{}

\item{extreme.usig}{}
}
\value{

}
\description{
Title
}
