% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition}
\alias{fractionalAirlineDecomposition}
\title{Title}
\usage{
fractionalAirlineDecomposition(
  y,
  period,
  adjust = T,
  sn = F,
  stde = F,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{y}{}

\item{period}{Period of the seasonal component}

\item{adjust}{True if an actual fractional airline model is used. False if the period is rounded to the nearest integer}

\item{sn}{Signal/noise decomposition. The signal is the seasonally adjusted series and the noise the seasonal component}

\item{stde}{True if standard deviations of the components must be computed. In some cases (memory limits), it is currently not possible to compute them}

\item{nbcasts}{Number of backcasts}

\item{nfcasts}{Number of forecasts}
}
\value{

}
\description{
Title
}
